Move file_langs into files_service
Relocate utils/file_langs.py into src/main_app/api_services/files_service and adapt the codebase accordingly. Introduce FileLanguagesMap dataclass to files_service.objects and export it; update files_service.__init__ exports. Update imports in FilesService (add get_file_languages wrapper), jobs worker, public routes, utils package, and tests to the new module path. Minor adjustments to FileLanguagesMap construction/return values were made to use the new dataclass defaults.
