+++ src/main_app/api_services/files_service/__init__.py
@@ -1,5 +1,6 @@
 """Utility modules for the main application."""
 
+from .file_langs import FileLanguagesMap, get_file_languages
 from .files_helpers import (
     get_file_info,
 )
@@ -21,4 +22,6 @@ __all__ = [
     "UploadService",
     "get_file_info",
     "UploadFile",
+    "get_file_languages",
+    "FileLanguagesMap",
 ]
