+++ tests/unit/jobs_workers/admin_jobs_workers/add_lang_categories_to_owid_pages/conftest.py
@@ -6,7 +6,7 @@ from unittest.mock import MagicMock
 
 import pytest
 
-from src.main_app.utils.file_langs import FileLanguagesMap
+from src.main_app.api_services.files_service.file_langs import FileLanguagesMap
 
 
 @pytest.fixture
