+++ tests/unit/jobs_workers/admin_jobs_workers/add_lang_categories_to_owid_pages/test_worker.py
@@ -7,12 +7,12 @@ from unittest.mock import MagicMock
 
 import pytest
 
+from src.main_app.api_services.files_service import FileLanguagesMap
 from src.main_app.jobs_workers.admin_jobs_workers.add_lang_categories_to_owid_pages.worker import (
     AddLangCategoriesWorker,
     PageInfo,
 )
 from src.main_app.jobs_workers.objects import JobsRunner
-from src.main_app.utils.file_langs import FileLanguagesMap
 
 
 @pytest.fixture
